Rename DateRange class as its own File
